// frontend/src/app/core/models/models.ts
export interface AdherentRequest {
  prenomAr: string;
  nomAr: string;
  categorie: string;
  grade: string;
  matriculeBR: string;
  matricule: string;
  dateNaissance: string;
  lieuNaissance: string;
  dateRadiation?: string | null;
  motifRadiation?: string | null;
  dateDeces?: string | null;
  causeDeces?: string | null;
  dernierUnite: string;
  formationUnite: string;
  telephone1: string;
  telephone2?: string | null;
  adresse: string;
  email: string;
  situationCategorie: string;
  pension: boolean;
  cin: string;
}

export interface AdherentResponse extends AdherentRequest {
  id: number;
}
// ...
export type MutuelleCourrierType = 'INTERNE' | 'EXTERNE';
export type MutuelleDossierType = 'ALD' | 'NORMAL' | 'DENTAIRE';

export interface MutuelleDossierRequest {
  adherentId: number;
  typeCourrier: MutuelleCourrierType;
  numeroOrdre: string;
  numeroEnvoi?: string | null;
  dateEnvoi?: string | null;
  dateReception?: string | null;
  designation: string;
  typeDossier: MutuelleDossierType;
  centreSoin?: string | null;
  observation?: string | null;
}